+++ R/pad.r
@@ -6,6 +6,8 @@
 #' @param width Minimum width of padded strings.
 #' @param side Side on which padding character is added (left, right or both).
 #' @param pad Single padding character (default is a space).
+#' @param use_length If `TRUE`, use the number of characters instead of the
+#'   total of character widths (see [stringi::stri_width]).
 #' @return A character vector.
 #' @seealso [str_trim()] to remove whitespace;
 #'   [str_trunc()] to decrease the maximum width of a string.
@@ -24,12 +26,12 @@
 #'
 #' # Longer strings are returned unchanged
 #' str_pad("hadley", 3)
-str_pad <- function(string, width, side = c("left", "right", "both"), pad = " ") {
+str_pad <- function(string, width, side = c("left", "right", "both"), pad = " ", use_length = FALSE) {
   side <- match.arg(side)
 
   switch(side,
-    left = stri_pad_left(string, width, pad = pad),
-    right = stri_pad_right(string, width, pad = pad),
-    both = stri_pad_both(string, width, pad = pad)
+    left = stri_pad_left(string, width, pad = pad, use_length = use_length),
+    right = stri_pad_right(string, width, pad = pad, use_length = use_length),
+    both = stri_pad_both(string, width, pad = pad, use_length = use_length)
   )
 }