+++ tests/testthat/test-pad.r
@@ -16,3 +16,11 @@ test_that("directions work for simple case", {
   expect_equal(pad("left"),   "       had")
   expect_equal(pad("both"),   "   had    ")
 })
+
+test_that("padding based of length works", {
+  # \u4e2d is a 2-characters-wide Chinese character
+  pad <- function(...) str_pad("\u4e2d", ..., side = "both")
+
+  expect_equal(pad(width = 6),                    "  \u4e2d  ")
+  expect_equal(pad(width = 5, use_length = TRUE), "  \u4e2d  ")
+})